refactored the game boostrap code and create Application interface.
