+++ src/main/java/net/forgottenschism/application/Application.java
@@ -1,0 +1,8 @@
+package net.forgottenschism.application;
+
+import org.newdawn.slick.Game;
+
+public interface Application extends Game
+{
+	void setApplicationBootstrap(ApplicationBootstrap applicationBootstrap);
+}
