+++ src/main/java/net/forgottenschism/main/Main.java
@@ -1,15 +1,12 @@
 package net.forgottenschism.main;
 
-import net.forgottenschism.constants.Constants;
-import org.newdawn.slick.CanvasGameContainer;
+import net.forgottenschism.application.ApplicationBootstrap;
+import net.forgottenschism.application.ForgottenSchismGame;
+
 import org.newdawn.slick.SlickException;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import javax.swing.*;
-import java.awt.event.WindowAdapter;
-import java.awt.event.WindowEvent;
-import java.io.File;
 
 public class Main
 {
@@ -17,8 +14,8 @@ public class Main
 
 	public static void main(String[] args) throws SlickException
 	{
-		GameApplicationBootstrap gameApplicationBootstrap = new GameApplicationBootstrap();
+		ApplicationBootstrap applicationBootstrap = new ApplicationBootstrap();
 
-		gameApplicationBootstrap.start();
+		applicationBootstrap.start(new ForgottenSchismGame());
 	}
 }
